requirements.txt
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fonttools==4.60.1
greenlet==3.2.4
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
msgpack==1.1.1
numpy==2.3.3
packaging==25.0
pandas==2.2.3
pillow==11.3.0
psycopg2-binary==2.9.11
pynvim==0.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
threadpoolctl==3.6.0
